requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9

# Channels for WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Utilities
PyJWT==2.8.0
python-dotenv==1.0.0

# For development
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
redis==5.0.1